fix: "*" routes can now be traversed down.
Assuming we had
```
{
	"*": {
		test: true
	}
}
```

The original functionality would be an acl of anything:test:true to work,
as its a wild card. IE if there are no string matches but there is a wild
card, pass through and match there for future checks.

This would stop at that point instead, causing a failure where it wasn't
expected.

Two tests were added to confirm this - and they failed as expected. Found
the culprit-  a missing can=true. Tests pass now.