fix: Sub "*" permissions were being passed too permissiviely.
An improperly formatted test would cause a test that checked for failure
to fail for unrelated reasons. The actual outcome was a "true", which
should have failed the test.

Added a bit more tests on sub checks under a "*" category, and fixed the
broken test.

Then fixed the cause of the broken test, allowing a "*" category to be
used in the middle of an ACL properly.